Verify: Consolidate download fullfile fallback to be reused
Moving the fullfile fallback to a single file and exporting it in
swupd.h in order to be reused on update and bundle add.

Tests:
 - Changing verify tests becase download function won't be called when
 there's no file to be downloaded
