+++ test/functional/verify/directory-tree-deleted/lines-checked
@@ -1,7 +1,5 @@
 Verifying version 10
 Verifying files
-Starting download of remaining update content. This may take a while...
-Finishing download of update content...
 Adding any missing files
 Fixing modified files
 REGEXP:Deleted .*/target-dir/testdir1/testdir2/testfile
