+++ test/functional/verify/picky-ghosted-missing/lines-checked
@@ -1,7 +1,5 @@
 Verifying version 10
 Verifying files
-Starting download of remaining update content. This may take a while...
-Finishing download of update content...
 Adding any missing files
 Fixing modified files
 REGEXP:--picky removing extra files under .*
